+++ src/main/java/org/spongepowered/api/event/SpongeEventFactory.java
@@ -26,7 +26,6 @@
 package org.spongepowered.api.event;
 
 import com.flowpowered.math.vector.Vector3d;
-import com.flowpowered.math.vector.Vector3f;
 import com.flowpowered.math.vector.Vector3i;
 import com.google.common.base.Optional;
 import com.google.common.cache.CacheBuilder;
@@ -91,26 +90,26 @@ import org.spongepowered.api.event.entity.EntityTameEvent;
 import org.spongepowered.api.event.entity.EntityTeleportEvent;
 import org.spongepowered.api.event.entity.EntityUpdateEvent;
 import org.spongepowered.api.event.entity.ProjectileLaunchEvent;
-import org.spongepowered.api.event.entity.living.player.PlayerBreakBlockEvent;
-import org.spongepowered.api.event.entity.living.player.PlayerChangeBlockEvent;
-import org.spongepowered.api.event.entity.living.player.PlayerChangeGameModeEvent;
-import org.spongepowered.api.event.entity.living.player.PlayerChangeWorldEvent;
-import org.spongepowered.api.event.entity.living.player.PlayerChatEvent;
-import org.spongepowered.api.event.entity.living.player.PlayerDeathEvent;
-import org.spongepowered.api.event.entity.living.player.PlayerDropItemEvent;
-import org.spongepowered.api.event.entity.living.player.PlayerHarvestBlockEvent;
-import org.spongepowered.api.event.entity.living.player.PlayerInteractBlockEvent;
-import org.spongepowered.api.event.entity.living.player.PlayerInteractEntityEvent;
-import org.spongepowered.api.event.entity.living.player.PlayerInteractEvent;
-import org.spongepowered.api.event.entity.living.player.PlayerJoinEvent;
-import org.spongepowered.api.event.entity.living.player.PlayerMoveEvent;
-import org.spongepowered.api.event.entity.living.player.PlayerPickUpItemEvent;
-import org.spongepowered.api.event.entity.living.player.PlayerPlaceBlockEvent;
-import org.spongepowered.api.event.entity.living.player.PlayerQuitEvent;
-import org.spongepowered.api.event.entity.living.player.PlayerUpdateEvent;
-import org.spongepowered.api.event.entity.living.player.fishing.PlayerCastFishingLineEvent;
-import org.spongepowered.api.event.entity.living.player.fishing.PlayerHookedEntityEvent;
-import org.spongepowered.api.event.entity.living.player.fishing.PlayerRetractFishingLineEvent;
+import org.spongepowered.api.event.entity.player.PlayerBreakBlockEvent;
+import org.spongepowered.api.event.entity.player.PlayerChangeBlockEvent;
+import org.spongepowered.api.event.entity.player.PlayerChangeGameModeEvent;
+import org.spongepowered.api.event.entity.player.PlayerChangeWorldEvent;
+import org.spongepowered.api.event.entity.player.PlayerChatEvent;
+import org.spongepowered.api.event.entity.player.PlayerDeathEvent;
+import org.spongepowered.api.event.entity.player.PlayerDropItemEvent;
+import org.spongepowered.api.event.entity.player.PlayerHarvestBlockEvent;
+import org.spongepowered.api.event.entity.player.PlayerInteractBlockEvent;
+import org.spongepowered.api.event.entity.player.PlayerInteractEntityEvent;
+import org.spongepowered.api.event.entity.player.PlayerInteractEvent;
+import org.spongepowered.api.event.entity.player.PlayerJoinEvent;
+import org.spongepowered.api.event.entity.player.PlayerMoveEvent;
+import org.spongepowered.api.event.entity.player.PlayerPickUpItemEvent;
+import org.spongepowered.api.event.entity.player.PlayerPlaceBlockEvent;
+import org.spongepowered.api.event.entity.player.PlayerQuitEvent;
+import org.spongepowered.api.event.entity.player.PlayerUpdateEvent;
+import org.spongepowered.api.event.entity.player.fishing.PlayerCastFishingLineEvent;
+import org.spongepowered.api.event.entity.player.fishing.PlayerHookedEntityEvent;
+import org.spongepowered.api.event.entity.player.fishing.PlayerRetractFishingLineEvent;
 import org.spongepowered.api.event.message.CommandEvent;
 import org.spongepowered.api.event.message.MessageEvent;
 import org.spongepowered.api.event.server.StatusPingEvent;
@@ -596,7 +595,7 @@ public final class SpongeEventFactory {
      * @return A new instance of the event
      */
     public static EntityDisplaceEvent createEntityDisplace(Game game, Entity entity,
-            Location oldLocation, Location newLocation, Vector3f rotation) {
+            Location oldLocation, Location newLocation, Vector3d rotation) {
         Map<String, Object> values = Maps.newHashMap();
         values.put("game", game);
         values.put("entity", entity);
@@ -721,7 +720,7 @@ public final class SpongeEventFactory {
      * @return A new instance of the event
      */
     public static EntityMoveEvent createEntityMove(Game game, Entity entity,
-            Location oldLocation, Location newLocation, Vector3f rotation) {
+            Location oldLocation, Location newLocation, Vector3d rotation) {
         Map<String, Object> values = Maps.newHashMap();
         values.put("game", game);
         values.put("entity", entity);
@@ -827,7 +826,7 @@ public final class SpongeEventFactory {
      * @return A new instance of the event
      */
     public static EntityTeleportEvent createEntityTeleport(Game game, Cause cause, Entity entity, Location oldLocation, Location newLocation,
-            Vector3f rotation, boolean keepsVelocity) {
+            Vector3d rotation, boolean keepsVelocity) {
         Map<String, Object> values = Maps.newHashMap();
         values.put("game", game);
         values.put("cause", Optional.fromNullable(cause));
@@ -1185,7 +1184,7 @@ public final class SpongeEventFactory {
      * @return A new instance of the event
      */
     public static PlayerInteractBlockEvent createPlayerInteractBlock(Game game, Cause cause, Player player, BlockLoc block,
-            EntityInteractionType interactionType, @Nullable Vector3f location) {
+            EntityInteractionType interactionType, @Nullable Vector3d location) {
         Map<String, Object> values = Maps.newHashMap();
         values.put("game", game);
         values.put("cause", Optional.fromNullable(cause));
@@ -1210,7 +1209,7 @@ public final class SpongeEventFactory {
      * @return A new instance of the event
      */
     public static PlayerInteractEntityEvent createPlayerInteractEntity(Game game, Player player, Entity targetEntity,
-            EntityInteractionType interactionType, @Nullable Vector3f location) {
+            EntityInteractionType interactionType, @Nullable Vector3d location) {
         Map<String, Object> values = Maps.newHashMap();
         values.put("game", game);
         values.put("entity", player);
@@ -1233,7 +1232,7 @@ public final class SpongeEventFactory {
      * @return A new instance of the event
      */
     public static PlayerInteractEvent createPlayerInteract(Game game, Player player, EntityInteractionType interactionType,
-            @Nullable Vector3f location) {
+            @Nullable Vector3d location) {
         Map<String, Object> values = Maps.newHashMap();
         values.put("game", game);
         values.put("entity", player);
@@ -1275,7 +1274,7 @@ public final class SpongeEventFactory {
      * @return A new instance of the event
      */
     public static PlayerMoveEvent createPlayerMove(Game game, Player player,
-            Location oldLocation, Location newLocation, Vector3f rotation) {
+            Location oldLocation, Location newLocation, Vector3d rotation) {
         Map<String, Object> values = Maps.newHashMap();
         values.put("game", game);
         values.put("entity", player);
